app/vibrantque-backend-main/requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
channels==4.0.0
channels-redis==4.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.14.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
msgpack==1.0.5
Pillow==10.0.0
promise==2.3
psycopg2==2.9.6
PyJWT==1.7.1
pytz==2023.3
redis==4.6.0
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
gunicorn==20.1.0
